requirements.txt
imageai==2.1.5
Pillow==8.0.1
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Uploads==0.2.1
Werkzeug==0.16.1
gevent==20.9.0
random
os
glob
base64